move bits/* includes into resp. sys/ files
With AT_* constants and inline syscalls, we are already leaking
a lot of stuff into includer's namespace. So let's leak the rest
as well and make it useful.

The idea is that if the code uses $syscall, it needs to know
all about $syscall's arguments -- types, constants and so on.
So ideally there should be a single #include <sys/$syscall.h>,
which would bring in everything needed to use $syscall.

There's only one case when this fails atm: ioctl.
Those need separate files for unrelated structs, so the caller
should include both <sys/ioctl.h> and the definitions needed
for the particular set of ctls in use.