svcmon: rework ring buffers, merge files
Buffers are now separated from svcrecs and moved closer
to poll fds. Past initial setup, the buffers are tied
to the fds but not to their parent recs.

Also, pollfd indexes are not passed around openly anymore,
only struct svcrec* which are then fed to recindex().

Also, major file re-arrangement.