wictl: drop SSID matching code
Doesn't look like it was a good idea. Doing it properly means also
doing it with the likes of -a/-b/-c and implementing a bunch of
dedicated commands in wimon. And BSSID argument without fixed-BSSID
mode is confusing at best. And chances of it all actually being ever
used are rather slim.