ifmon: internal dhcp (WIP)
Just like with EAPOL and the scanning code, it makes lots of sense
to keep dhcp close to the link monitoring code. Timeout tracking
becomes much easier, there's some chances of handling dns settings
nicely and so on.

This commit enables basic addr setting but the code is not yet complete.