wifi: de-couple from ifmon, wsupp no-device startup
The point of this change is to let wsupp start as a regular service
under `super`, and allow late device configuration via its control
interface. This way, ifmon would not need to start and supervise wsupp,
and itstead would only need to do a short IPC call (with a short-living
script) to let wsupp know which device to use.

This commit is breaking until ifmon is re-written to run event scripts
instead of trying to spawn wsupp.