wsupp: cache IEs and pass them to the client
This way, IE parsing code in the resident part can be kept to the
absolute minimum while also allowing the client to show as much as
possible about the APs to the user.

IEs are central to AP selection so doing the about brought lots of
changes both in wsupp and in the client.