lib: pointer-only qsort
The original interface for qsort was that of the POSIX libc, so that
the function could sort entries of any size. This project only really
needs to sort arrays of pointers.

Restricting the function to pointers only simplifies the code, and also
makes comparators much easier to write as they get their pointer by
value and not be reference (as in, void** to be deref'ed into void*).