msh: proxy signals for running children
Without this, SIGINT while msh is busy with `run foo` would kill msh
but leave the running process orphaned, an unexpected and mostly
undesirable behavior. At the cost of two extra syscalls, we can do
much better.

This also allows doing `run foo` in place of `exec foo` when spawning
processes with a decoupling msh script. The parent would send a signal
to msh, msh would pass it to the running process, and it would be
handled the same way it would if the process were exec'ed into instead.