msh: drop symbolic user/group names
Only allow integer IDs.

The dependency on /etc/passwd and /etc/group, is becoming a huge problem
with the possible move to a dedicated prefix, and the need for msh to be
usable from initrd.

It does not look like there are any good options for symbolic names,
and given the way msh is intended to be used, it should not be that
big of an issue, let's limit it to numeric IDs.