move stuff to /base
Configuration for this project is now in /base/etc, and
the binaries are meant to be installed in /base/bin, /base/sys.

Basically use a dedicated prefix for the project.

It is not really configuration at this point, the only way
to change it is to patch the sources, so let's stop pretending
it is, and also remove relevant keys from configure.